Task # 36
Added a range options GUI button through html to give the user the option to filter out lodging based on the distance from their current location. Once selected and typed into the range options button, when the user clicks on the hotel options GUI button, it will search nearby based on that distance.